Add files via upload
This is a DBMS mini project with fully functional front end and back end. I developed backend with SQL, JSP and Apache Tomcat as web server. The project can be used in real world smartphone stores for efficient querying and searches. Front end of this project is very simple and attractive.